made ready for merge
put some files in src directory. Deleted unnecessary files. changed file names to adhere to code agreement!
