+++ src/assets/level-assets/obstacles/easy/easy
