+++ src/assets/level-assets/obstacles/hard/hard
