+++ src/assets/level-assets/obstacles/medium/medium
