+++ src/js/players/player-1.js
@@ -1,5 +1,5 @@
 import { Actor, Vector, Input } from "excalibur";
-import { Resources, ResourceLoader } from "/js/resources.js";
+import { Resources, ResourceLoader } from "/src/js/resources.js";
 
 export class Taxi extends Actor {
   constructor() {
