+++ src/js/resources.js
